client/requirements.txt
kivy
python-socketio
